requirements/requirements.txt
transformers==4.9.2
sentence-transformers==2.0.0
faiss-cpu==1.7.1
numpy==1.21.2
requests==2.26.0
-e